#16 Added MsBuild and Gulp Build of the Esf.sln to the deployment pipeline. Full dev deployment from source is complete.

+++ source/Esf.DataAccess.Deployment/DeleteDb.ps1
@@ -1,2 +1,5 @@
 $dbDeploymentConfig = (Get-Content .\dbDeployment.config.json) -join "`n" | ConvertFrom-Json
-Remove-Item $dbDeploymentConfig.esFiddleDbPath -recurse
+if (Test-Path $dbDeploymentConfig.esFiddleDbPath)
+{
+	Remove-Item $dbDeploymentConfig.esFiddleDbPath -recurse
+}
